Refactor. js삽입방식 변경 및 모듈파일 함수화
