+++ js/api/api.js
@@ -1,4 +1,4 @@
-import config from "../config/config.js";
+import config from "../../config/config.js";
 const API_URL = `https://omdbapi.com/?apikey=${config.OMDB_API_KEY}`;
 const JSON_URL = "/data/movieData.json";
