+++ pages/bookmark.html
@@ -1,26 +1,24 @@
 <!DOCTYPE html>
 <html lang="en">
+  <head>
+    <meta charset="UTF-8" />
+    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
+    <title>Movie POP</title>
+    <script src="https://cdnjs.cloudflare.com/ajax/libs/bodymovin/5.10.2/lottie.min.js"></script>
+    <link rel="icon" href="/images/favicon.png" />
+    <link rel="stylesheet" href="/css/index.css" />
+  </head>
 
-<head>
-  <meta charset="UTF-8" />
-  <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-  <title>Movie POP</title>
-  <script src="https://cdnjs.cloudflare.com/ajax/libs/bodymovin/5.10.2/lottie.min.js"></script>
-  <link rel="icon" href="/images/favicon.png" />
-  <link rel="stylesheet" href="/css/index.css" />
-</head>
-
-<body>
-  <div class="header__wrapper"></div>
-  <div class="loading__wrapper">
-    <div class="loading"></div>
-  </div>
-  <section class="bookmark">
-    <h2 class="bookmark__title">BOOKMARKED</h2>
-    <div class="bookmark__content"></div>
-  </section>
-</body>
-<footer class="footer__wrapper"></footer>
-<script type="module" src="/js/bookmark.js"></script>
-
+  <body>
+    <div class="header__wrapper"></div>
+    <div class="loading__wrapper">
+      <div class="loading"></div>
+    </div>
+    <section class="bookmark">
+      <h2 class="bookmark__title">BOOKMARKED</h2>
+      <div class="bookmark__content"></div>
+    </section>
+  </body>
+  <footer class="footer__wrapper"></footer>
+  <script type="module" src="/js/index.js"></script>
 </html>
