+++ pages/detail.html
@@ -1,36 +1,34 @@
 <!DOCTYPE html>
 <html lang="en">
+  <head>
+    <meta charset="UTF-8" />
+    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
+    <link rel="icon" href="/images/favicon.png" />
+    <link rel="stylesheet" href="/css/index.css" />
+    <script src="https://cdnjs.cloudflare.com/ajax/libs/bodymovin/5.10.2/lottie.min.js"></script>
+    <script src="https://developers.kakao.com/sdk/js/kakao.js"></script>
+    <title>Movie POP</title>
+  </head>
 
-<head>
-  <meta charset="UTF-8" />
-  <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-  <link rel="icon" href="/images/favicon.png" />
-  <link rel="stylesheet" href="/css/index.css" />
-  <script src="https://cdnjs.cloudflare.com/ajax/libs/bodymovin/5.10.2/lottie.min.js"></script>
-  <script src="https://developers.kakao.com/sdk/js/kakao.js"></script>
-  <title>Movie POP</title>
-</head>
+  <body>
+    <div class="header__wrapper"></div>
+    <div class="loading__wrapper"></div>
+    <main class="detail__container">
+      <div class="loading"></div>
+      <button type="button" class="btn-goback">
+        <span class="a11y-hidden">go back</span>
+      </button>
+      <button type="button" class="btn-bookmark not-bookmarked">
+        <span class="a11y-hidden">book mark</span>
+      </button>
 
-<body>
-  <div class="header__wrapper"></div>
-  <div class="loading__wrapper"></div>
-  <main class="detail__container">
-    <div class="loading"></div>
-    <button type="button" class="btn-goback">
-      <span class="a11y-hidden">go back</span>
-    </button>
-    <button type="button" class="btn-bookmark not-bookmarked">
-      <span class="a11y-hidden">book mark</span>
-    </button>
+      <button type="button" class="btn-share">
+        <span class="a11y-hidden">share</span>
+      </button>
 
-    <button type="button" class="btn-share">
-      <span class="a11y-hidden">share</span>
-    </button>
-
-    <div class="detail__movie-info-wrapper"></div>
-  </main>
-  <footer class="footer__wrapper"></footer>
-  <script src="/js/detail.js" type="module"></script>
-</body>
-
+      <div class="detail__movie-info-wrapper"></div>
+    </main>
+    <footer class="footer__wrapper"></footer>
+    <script type="module" src="/js/index.js"></script>
+  </body>
 </html>
